worker.js - Fix for saving before init() had run. Should stop the data flushing issue... worker_ca_monster.js - Some more fixes for taint.

+++ _normal.user.js
@@ -3,7 +3,7 @@
 // @namespace	golem
 // @description	Auto player for Castle Age on Facebook. If there's anything you'd like it to do, just ask...
 // @license		GNU Lesser General Public License; http://www.gnu.org/licenses/lgpl.html
-// @version		31.5.1077
+// @version		31.5.1078
 // @include		http://apps.facebook.com/castle_age/*
 // @include		https://apps.facebook.com/castle_age/*
 // @require		http://cloutman.com/jquery-1.4.2.min.js
@@ -27,7 +27,7 @@ var isRelease = false;
 var script_started = Date.now();
 // Version of the script
 var version = "31.5";
-var revision = 1077;
+var revision = 1078;
 // Automatically filled from Worker:Main
 var userID, imagepath, APP, APPID, APPNAME, PREFIX; // All set from Worker:Main
 // Detect browser - this is rough detection, mainly for updates - may use jQuery detection at a later point
@@ -784,7 +784,7 @@ NOTE: If there is a work() but no display() then work(false) will be called befo
 ._load(type)			- Loads data / option from storage, merges with current values, calls .update(type) on change
 ._save(type)			- Saves data / option to storage, calls .update(type) on change
 
-._flush()				- Calls this._save() then deletes this.data if !this.settings.keep
+._flush()				- Calls this._save() then deletes this.data if !this.settings.keep ** PRIVATE **
 ._unflush()				- Loads .data if it's not there already
 
 ._parse(change)			- Calls this.parse(change) inside a try / catch block
@@ -896,12 +896,8 @@ Worker.flush = function() {
 		Workers[i]._update(null, 'run');
 	}
 	for (i in Workers) {
-//		console.log('Worker.flush(): '+i+'._save()');
-		Workers[i]._save();
-		if (!Workers[i].settings.keep) {
-//			console.log('Worker.flush(): delete '+i+'.data');
-			delete Workers[i]['data'];
-		}
+//		console.log('Worker.flush(): '+i+'._flush()');
+		Workers[i]._flush();
 	}
 };
 
@@ -916,10 +912,10 @@ Worker._triggers_ = [];// Used for this._trigger
  * @protected
  */
 Worker.prototype._flush = function() {
-	if (this['data']) {
+	if (this._loaded) {
 		this._pushStack();
-		this._save('data');
-		if (!this.settings.keep) {
+		this._save();
+		if (this['data'] && !this.settings.keep) {
 			delete this['data'];
 		}
 		this._popStack();
@@ -1305,7 +1301,7 @@ Worker.prototype._save = function(type) {
 		}
 		return n;
 	}
-	if (!this._datatypes[type] || this._saving[type] || (this.settings.taint && !this._taint[type]) || this[type] === undefined || this[type] === null) {
+	if (!this._datatypes[type] || this._saving[type] || this[type] === undefined || this[type] === null || (this.settings.taint && !this._taint[type])) {
 		return false;
 	}
 	this._saving[type] = true;
@@ -1318,7 +1314,7 @@ Worker.prototype._save = function(type) {
 		return false; // exit so we don't try to save mangled data over good data
 	}
 	n = (this._rootpath ? userID + '.' : '') + type + '.' + this.name;
-	if (this.settings.taint || this._taint[type] || getItem(n) !== v) { // First two are to save the extra getItem from being called
+	if (this._taint[type] || getItem(n) !== v) { // First two are to save the extra getItem from being called
 		this._pushStack();
 		this._taint[type] = false;
 		this._timestamps[type] = Date.now();
@@ -1591,29 +1587,32 @@ Worker.prototype._update = function(event, type) {
 	if (this._loaded) {
 		this._pushStack();
 		var i, done = false, events;
-		if (isString(event)) {
-			event = {type:event};
-		} else if (!isObject(event)) {
-			event = {};
+		if (event) {
+			if (isString(event)) {
+				event = {type:event};
+			} else if (!isObject(event)) {
+				event = {};
+			}
+			if (event.type && (isFunction(this.update) || isFunction(this['update_'+event.type]))) {
+				event.worker = isWorker(event.worker) ? event.worker.name : event.worker || this.name;
+				if (type !== 'purge' && (i = this._updates_.findEvent(event.worker, event.type, event.id)) >= 0) { // Delete from update queue
+					this._updates_.splice(i,1);
+				}
+				if (type !== 'add' && type !== 'delete') { // Add to update queue, old key already deleted
+					this._updates_.unshift($.extend({}, event));
+				}
+				if (type === 'purge') { // Purge the update queue immediately - don't do anything with the entries
+					this._updates_ = [];
+				}
+				if (this._updates_.length) {
+					Worker.updates[this.name] = true;
+				} else {
+					delete Worker.updates[this.name];
+				}
+			}
 		}
-		if (event.type && (isFunction(this.update) || isFunction(this['update_'+event.type]))) {
-			event.worker = isWorker(event.worker) ? event.worker.name : event.worker || this.name;
-			if (type !== 'purge' && (i = this._updates_.findEvent(event.worker, event.type, event.id)) >= 0) { // Delete from update queue
-				this._updates_.splice(i,1);
-			}
-			if (type !== 'add' && type !== 'delete') { // Add to update queue, old key already deleted
-				this._updates_.unshift($.extend({}, event));
-				Worker.updates[this.name] = true;
-			}
-			if (type === 'purge') { // Purge the update queue immediately - don't do anything with the entries
-				this._updates_ = [];
-				delete Worker.updates[this.name];
-			}
-		}
-		if (type === 'run' && this._updates_.length) { // Go through the event list and process each one
-			if (isUndefined(this.data) && this._datatypes.data) {
-				this._unflush();
-			}
+		if (type === 'run' && Worker.updates[this.name]) { // Go through the event list and process each one
+			this._unflush();
 			events = this._updates_;
 			this._updates_ = [];
 			while (events.length && done !== true) {
@@ -9940,7 +9939,7 @@ Monster.parse = function(change) {
 		}
 		mid = uid+'_' + (Page.page === 'festival_battle_monster' ? 'f' : (types[i].mpool || 4));
 		if (this.runtime.check === mid) {
-			this.runtime.check = false;
+			this.set(['runtime','check'], false);
 		}
 		monster = data[mid] = data[mid] || {};
 		monster.type = type_label;
@@ -10118,17 +10117,18 @@ Monster.parse = function(change) {
 			monster.total = Math.ceil(100 * sum(monster.damage) / (monster.health === 100 ? 0.1 : (100 - monster.health)));
 		}
 		monster.eta = now + (Math.floor((monster.total - sum(monster.damage)) / monster.dps) * 1000);
+		this._taint[data] = true;
 //		this.runtime.used.stamina = 0;
 //		this.runtime.used.energy = 0;
 	} else if (Page.page === 'monster_dead' || $('div[style*="no_monster_back.jpg"]').length) {
 		console.log(warn('Found a timed out monster.'));
 		if (clicked) {
 			console.log(warn(), 'Deleting ' + data[this.runtime.mid].name + "'s " + data[this.runtime.mid].type);
-			delete data[this.runtime.mid];
+			this.set(['data',this.runtime.mid]);
 		} else {
 			console.log(warn('Unknown monster (timed out)'));
 		}
-		this.runtime.check = false;
+		this.set(['runtime','check'], false);
 		return false;
 /*	} else if (['monster_monster_list', 'battle_raid', 'festival_monster_list'].indexOf(Page.page)>=0) { // Check monster / raid list
 				switch($children.eq(2).find('input[type="image"]').attr('src').regex(/(\w+)\.(gif|jpg)/)[0]) {
